fix(cpu): add platform-specific stubs with collect_cpu implementation
Add macOS and Windows CPU collector stubs with proper collect_cpu() method.
Update CPU module to use per-platform implementations matching network collector pattern.
Log warnings when CPU tracking is unavailable on non-Linux platforms.
